requirements.txt
fastapi
uvicorn
python-dotenv
pillow
pdf2image
google-generativeai
